Update Get-DatabricksDBFSFile.ps1 (#107)
* Update Get-DatabricksDBFSFile.ps1

I can't see exactly why you added the PS version required to 6 or higher. But this provided with some errors inside PS version 5.1. 
reference:  https://github.com/PowerShell/PowerShell/issues/3131
Because I run this script inside Azure DevOps pipelines (which has no possibility to use PS 6 or higher).
I changed the script to handle the commands and rest call to capture the file even with PS 5.1.
If you find another way around this no problem, just pointing out the problem and where the workaround should be handled.
*Also remember your scripts are being used by people in automated flaws, changing them will have some impact.

* Update Get-DatabricksDBFSFile.ps1

deleted the extra  S in the function that was a typo